+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to semantic versioning.
 
 ## 0.7.0 (_Unreleased_)
 
+- Add `aperture` function to the `Array` module.
 - Provide the input value to the fallback function of `guard`.
 
 ## 0.6.0 (2020-11-20)